+++ library/ruler-jvm/src/main/java/io/mehow/ruler/InRageUnitFitter.kt
@@ -1,0 +1,13 @@
+package io.mehow.ruler
+
+/**
+ * Selects a unit based on a best fit of a [distance][Length.distance] in [unit bounds][LengthUnit.bounds].
+ */
+public object InRageUnitFitter : UnitFitter {
+  override fun <T : LengthUnit<T>> findFit(
+    units: Iterable<T>,
+    length: Length<T>,
+  ): T? = length.distance.abs().let { distance ->
+    units.firstOrNull { distance in it }
+  }
+}