+++ library/ruler-jvm/src/test/java/io/mehow/ruler/test/DistanceGenerator.kt
@@ -4,14 +4,20 @@ import io.kotest.property.Arb
 import io.kotest.property.RandomSource
 import io.kotest.property.Sample
 import io.mehow.ruler.Distance
+import io.mehow.ruler.LengthUnit
 
 internal class DistanceGenerator private constructor(
   private val range: ClosedRange<Distance>,
 ) : Arb<Distance>() {
-  override fun edgecases() = (setOf(range.start, range.endInclusive) + when (Distance.Zero) {
-    in range -> setOf(Distance.Zero)
-    else -> emptySet()
-  }).toList()
+  private val edgeCases = listOfNotNull(
+      range.start,
+      range.start + Distance.Epsilon,
+      Distance.Zero,
+      range.endInclusive - Distance.Epsilon,
+      range.endInclusive,
+  ).filter { it in range }.distinct()
+
+  override fun edgecases() = edgeCases
 
   override fun sample(rs: RandomSource) = rs.random
       .nextDistance(range)
@@ -22,5 +28,7 @@ internal class DistanceGenerator private constructor(
       min: Distance = Distance.Min,
       max: Distance = Distance.Max,
     ) = DistanceGenerator(min..max)
+
+    fun create(unit: LengthUnit<*>) = DistanceGenerator(unit.bounds)
   }
 }