+++ library/ruler-jvm/src/test/java/io/mehow/ruler/test/LengthUtils.kt
@@ -1,0 +1,5 @@
+package io.mehow.ruler.test
+
+import io.mehow.ruler.LengthUnit
+
+internal val LengthUnit<*>.capacity get() = bounds.endInclusive.toLength(this).measure.toLong()