+++ docs/changelog.md
@@ -8,9 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `Distance.Epsilon` constant that represents smallest possible change in of `Distance`.
+- `LengthUnits.bounds` property to the public API. It holds min and max distances that only that unit can represent as a natural count of self.
+- `Length.withFittingUnit(units, fitter)` that selects the best unit for a length using supplied units and fitting algorithm. A complementary `UnitFitter` interface is added along with `InRangeUnitFitter` and `LogDistanceUnitFitter` implementations.
 
 ### Changed
-- Lengths with zero distance to not change their units when `Length.withAutoUnit()` is used.
 - Added one nanometer to `Distance.Min`. This change was made in order to make `Distance.abs()` operation safe.
 
 ## [1.0.0] - 2020-12-28